Bucketload of changes to make numbers a first-class set type.

+++ src/test/java/testdata/MyTestData.java
@@ -1,7 +1,7 @@
 package testdata;
 
 public class MyTestData {
-	public static int test1(short incoming) {
+	public static int test1(int incoming) {
 		if (incoming < 3) {
 			return 2;
 		} else {
@@ -9,14 +9,14 @@ public class MyTestData {
 		}
 	}
 	
-	public static int test2(short incoming1) {
+	public static int test2(int incoming1) {
 		int foo = 0;
 		int bar = incoming1;
 		int incoming2 = incoming1 * 2;
 		
 		if (incoming1 > 2) {
 			incoming2 = incoming1 * incoming2;
-			incoming1 = (short) (0 - incoming1);
+			incoming1 = 0 - incoming1;
 			bar += incoming1;
 			foo += 3;
 		}
